make avatar component cleaner
